Bugfix cjson:metadata handling when using wdio V7 (#67)
* Renamed property cjson_metadata to cjson:metadata

The W3C webdriver protocol specifies a list of standard capabilities
(https://www.w3.org/TR/webdriver1/#capabilities).
Everything that's not a standard capability is considered an extension
capability. Extension capabilities' keys must contain a colon character
to separate an implementation specific namespace from the name of the
capability itself. Thus, the capability key cjson_metadata is not valid
and will be rejected by a remote end that adheres to the specification.

It looks like previous versions of wdio transformed the capability key
cjson:metadata to cjson_metadata, but at least with wdio 7.12.2 this
seems not to be the case anymore. Thus, the property cjson_metadata of
W3CCapabilitiesExtended been renamed to cjson:metadata in the code.

* Updated README.md to state that cjson:metadata works with wdio V7

Co-authored-by: Martin Hanysz <martin.hanysz@nexenio.com>